clean_count to utils
