+++ peregrine/resources/submission/graphql/counts.py
@@ -1,4 +1,3 @@
-import sqlalchemy as sa
 from .schema import SchemaQuery
 from psqlgraph import Node
 import base
@@ -13,20 +12,13 @@ from .base import (
     munge,
 )
 
+from .util import (
+    clean_count,
+)
+
 from gdcgraphql import (
     Query,
 )
-
-def clean_count(q):
-    """Returns the count from this query without pulling all the columns
-    This gets the count from a query without doing a subquery
-    The subquery would pull all the information from the DB
-    and cause statement timeouts with large numbers of rows
-    Args:
-        q (psqlgraph.query.GraphQuery): The current query object.
-    """
-    query_count = q.options(sa.orm.lazyload('*')).statement.with_only_columns([sa.func.count()]).order_by(None)
-    return q.session.execute(query_count).scalar()
 
 class NodeCountQuery(base.GraphQLQuery):
